app/projects/urls.py
from django.urls import path
from .views import ProjectListView, ProjectDetailView, ProjectCreateView

app_name = "projects"

urlpatterns = [
    path("", ProjectListView.as_view(), name="list"),
    path("create/", ProjectCreateView.as_view(), name="create"),
    path("<int:pk>/", ProjectDetailView.as_view(), name="detail"),
]