requirements/dev.txt
-r base.txt
ruff
pytest
pytest-django
